+++ CHANGES.md
@@ -9,6 +9,7 @@
 * added ability to enter Bitbucket credentials to Install tabs if not already present
 * added a couple of `wp_cache_flush()` calls to ensure extra headers load
 * moved action/filter hook calls out of constructors, make @carlalexander happy
+* refactored to incorporate GitLab personal access tokens
 
 #### 5.6.2 / 2016-09-24
 * added reset of _update\_plugins_ and _update\_themes_ transient with _Refresh Transients_